workload/schemachange: extract opType into a new file
Previously, `opType` and static metadata structures where located in the
operation_generator.go file. Due to the massive size and of the file and
mixture of re possibilities within it, I found myself getting lost quite
frequently.

This commit extracts `opType` and said metadata structures into a
dedicated `optype.go` which better matches the generated
`optype_string.go` file.

In addition to the extraction, this commit performs a few quality of
life changes:
* All runtime assertions have been combined into a single init function
* Runtime assertions now have friendlier messages
* `opDeclarative` has been removed as it is implied by
	`opDeclarativeVersion`
* Add a handful of comments for future maintainers

Epic: CRDB-3265
Release note: None
